+++ pages/index.js
@@ -1,16 +1,16 @@
 import React from 'react'
 import Link from 'next/link'
-import withRoot from '../components/Layout'
+import Layout from '../components/Layout'
 
 const IndexPage = () => (
-  <div>
+  <Layout>
     <h1>Hi people</h1>
     <p>Welcome to your new Gatsby site.</p>
     <p>Now go build something great.</p>
     <Link href={'/page'}>
       <a>Go to page 2</a>
     </Link>
-  </div>
+  </Layout>
 )
 
-export default withRoot(IndexPage)
+export default IndexPage
