+++ pages/page.js
@@ -1,15 +1,15 @@
 import React from 'react'
 import Link from 'next/link'
-import withRoot from '../components/Layout'
+import Layout from '../components/Layout'
 
 const SecondPage = () => (
-  <div>
+  <Layout>
     <h1>Hi from the second page</h1>
     <p>Welcome to age 2</p>
     <Link href={'/'}>
       <a>Go back to the homepage</a>
     </Link>
-  </div>
+  </Layout>
 )
 
-export default withRoot(SecondPage)
+export default SecondPage
